+++ library-gen/Stratosphere/Resources.hs
@@ -179,6 +179,7 @@ import Stratosphere.Resources.ElasticBeanstalkConfigurationTemplate as X
 import Stratosphere.Resources.ElasticBeanstalkEnvironment as X
 import Stratosphere.Resources.ElasticLoadBalancingLoadBalancer as X
 import Stratosphere.Resources.ElasticLoadBalancingV2Listener as X
+import Stratosphere.Resources.ElasticLoadBalancingV2ListenerCertificateResource as X
 import Stratosphere.Resources.ElasticLoadBalancingV2ListenerRule as X
 import Stratosphere.Resources.ElasticLoadBalancingV2LoadBalancer as X
 import Stratosphere.Resources.ElasticLoadBalancingV2TargetGroup as X
@@ -958,6 +959,7 @@ data ResourceProperties
   | ElasticBeanstalkEnvironmentProperties ElasticBeanstalkEnvironment
   | ElasticLoadBalancingLoadBalancerProperties ElasticLoadBalancingLoadBalancer
   | ElasticLoadBalancingV2ListenerProperties ElasticLoadBalancingV2Listener
+  | ElasticLoadBalancingV2ListenerCertificateResourceProperties ElasticLoadBalancingV2ListenerCertificateResource
   | ElasticLoadBalancingV2ListenerRuleProperties ElasticLoadBalancingV2ListenerRule
   | ElasticLoadBalancingV2LoadBalancerProperties ElasticLoadBalancingV2LoadBalancer
   | ElasticLoadBalancingV2TargetGroupProperties ElasticLoadBalancingV2TargetGroup
@@ -1400,6 +1402,8 @@ resourcePropertiesJSON (ElasticLoadBalancingLoadBalancerProperties x) =
   [ "Type" .= ("AWS::ElasticLoadBalancing::LoadBalancer" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (ElasticLoadBalancingV2ListenerProperties x) =
   [ "Type" .= ("AWS::ElasticLoadBalancingV2::Listener" :: String), "Properties" .= toJSON x]
+resourcePropertiesJSON (ElasticLoadBalancingV2ListenerCertificateResourceProperties x) =
+  [ "Type" .= ("AWS::ElasticLoadBalancingV2::ListenerCertificate" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (ElasticLoadBalancingV2ListenerRuleProperties x) =
   [ "Type" .= ("AWS::ElasticLoadBalancingV2::ListenerRule" :: String), "Properties" .= toJSON x]
 resourcePropertiesJSON (ElasticLoadBalancingV2LoadBalancerProperties x) =
@@ -1784,6 +1788,7 @@ resourceFromJSON n o =
          "AWS::ElasticBeanstalk::Environment" -> ElasticBeanstalkEnvironmentProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancing::LoadBalancer" -> ElasticLoadBalancingLoadBalancerProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancingV2::Listener" -> ElasticLoadBalancingV2ListenerProperties <$> (o .: "Properties")
+         "AWS::ElasticLoadBalancingV2::ListenerCertificate" -> ElasticLoadBalancingV2ListenerCertificateResourceProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancingV2::ListenerRule" -> ElasticLoadBalancingV2ListenerRuleProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancingV2::LoadBalancer" -> ElasticLoadBalancingV2LoadBalancerProperties <$> (o .: "Properties")
          "AWS::ElasticLoadBalancingV2::TargetGroup" -> ElasticLoadBalancingV2TargetGroupProperties <$> (o .: "Properties")